+++ lib/ps/windows_firewall/ps-bridge-ipsec.ps1
@@ -183,8 +183,11 @@ function create {
 
 function update {
     write-host "Updating $($Name)..."
+
+    # rules containing square brackets need to be escaped or nothing will match
+    $Name = $name.replace(']', '`]').replace('[', '`[')
+
     $params = @{
-        Name        = $Name;
     }
     if ($DisplayName) {
         $params.Add("NewDisplayName", $DisplayName)
@@ -282,8 +285,8 @@ function update {
         }
     }
 
-    if (Get-NetIPSecRule -Name $name -erroraction 'silentlycontinue') {
-        Set-NetIPSecRule @params -ErrorAction Stop
+    if (Get-NetIPSecRule -Name $name -erroraction SilentlyContinue) {
+        Set-NetIPSecRule -Name $name @params -ErrorAction Stop
     }
     else {
         throw "We were told to update firewall rule '$($name)' but it does not exist"
@@ -294,23 +297,14 @@ function delete {
     write-host "Deleting $($Name)..."
 
     # rules containing square brackets need to be escaped or nothing will match
-    # eg: "Ruby interpreter (CUI) 2.4.3p205 [x64-mingw32]"
     $Name = $name.replace(']', '`]').replace('[', '`[')
 
-    # Depending how rule was parsed (netsh vs ps) `$Name` will contain either
-    # `DisplayName` or rule ID. Therefore, delete by Displayname first, if this
-    # fails, fallback to `Name` and if this also fails, error the script
-    # (`-ErrorAction Stop`)
-    if (Get-NetIPSecRule -DisplayName $name -erroraction 'silentlycontinue') {
-        remove-NetIPSecRule -DisplayName $Name
-    }
-    elseif (Get-NetIPSecRule -Name $name -erroraction 'silentlycontinue') {
-        remove-NetIPSecRule -Name $Name -ErrorAction Stop
+    if (Get-NetIPSecRule -Name $name -ErrorAction SilentlyContinue) {
+        Remove-NetIPSecRule -Name $Name -ErrorAction Stop
     }
     else {
         throw "We were told to delete firewall rule '$($name)' but it does not exist"
     }
-
 }
 
 switch ($Target) {