+++ lib/ps/windows_firewall/ps-bridge.ps1
@@ -106,23 +106,14 @@ function delete {
     write-host "Deleting $($Name)..."
 
     # rules containing square brackets need to be escaped or nothing will match
-    # eg: "Ruby interpreter (CUI) 2.4.3p205 [x64-mingw32]"
     $Name = $name.replace(']', '`]').replace('[', '`[')
 
-    # Depending how rule was parsed (netsh vs ps) `$Name` will contain either
-    # `DisplayName` or rule ID. Therefore, delete by Displayname first, if this
-    # fails, fallback to `Name` and if this also fails, error the script
-    # (`-ErrorAction Stop`)
-    if (Get-NetFirewallRule -DisplayName $name -erroraction 'silentlycontinue') {
-        remove-netfirewallrule -DisplayName $Name
-    }
-    elseif (Get-NetFirewallRule -Name $name -erroraction 'silentlycontinue') {
-        remove-netfirewallrule -Name $Name -ErrorAction Stop
+    if (Get-NetFirewallRule -Name $name -ErrorAction SilentlyContinue) {
+        Remove-NetFirewallRule -Name $Name -ErrorAction Stop
     }
     else {
         throw "We were told to delete firewall rule '$($name)' but it does not exist"
     }
-
 }
 
 
@@ -224,8 +215,11 @@ function create {
 
 function update {
     write-host "Updating $($Name)..."
+
+    # rules containing square brackets need to be escaped or nothing will match
+    $Name = $name.replace(']', '`]').replace('[', '`[')
+
     $params = @{
-        Name = $Name;
     }
     if ($DisplayName) {
         $params.Add("NewDisplayName", $DisplayName)
@@ -322,8 +316,8 @@ function update {
         $params.Add("RemoteMachine", $RemoteMachine)
     }
 
-    if (Get-NetFirewallRule -Name $name -erroraction 'silentlycontinue') {
-        Set-NetFirewallRule @params -ErrorAction Stop
+    if (Get-NetFirewallRule -Name $name -ErrorAction SilentlyContinue) {
+        Set-NetFirewallRule -Name $name @params -ErrorAction Stop
     }
     else {
         throw "We were told to update firewall rule '$($name)' but it does not exist"